Move prepare_namespace from utils to namespace extension
